pyproject.toml
[project]
name = "graphrag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "neo4j-graphrag[openai]>=1.10.0",
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
